+++ inc/serializeproto/ParserProto.h
@@ -46,12 +46,12 @@ enum WireType
 class ParserProto
 {
 public:
-    ParserProto(IParserVisitor& visitor, const char* ptr, int size);
+    ParserProto(IParserVisitor& visitor, const char* ptr, ssize_t size);
 
     bool parseStruct(const std::string& typeName);
 
 private:
-    bool parseString(const char*& buffer, int& size);
+    bool parseString(const char*& buffer, ssize_t& size);
     void parseStructWire(const MetaField& field);
     bool parseStructIntern(const MetaStruct& stru);
     void parseArrayStruct(const MetaField& field);
