+++ inc/variant/VariantValueStruct.h
@@ -58,7 +58,7 @@ private:
     virtual bool add(const Variant& variant) override;
     virtual bool add(Variant&& variant) override;
     virtual ssize_t size() const override;
-    virtual void visit(IVariantVisitor& visitor, Variant& variant, int index, int level, int size, const std::string& name) override;
+    virtual void visit(IVariantVisitor& visitor, Variant& variant, ssize_t index, int level, ssize_t size, const std::string& name) override;
 
 
     VariantStruct::iterator find(const std::string& name);
